+++ app/src/main/AndroidManifest.xml
@@ -11,12 +11,8 @@
         android:supportsRtl="true"
         android:theme="@style/Theme.GOSOPTAndroid"
         tools:targetApi="31">
-        <activity
-            android:name=".MainActivity"
-            android:windowSoftInputMode="adjustPan"
-            android:exported="true">
-        </activity>
-        <activity android:name=".Main2Activity"
+
+        <activity android:name=".MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
